Replace the pendingStops set with a couple flags in the Thread object
The main idea here is make the "wait for stops" part of
GdbServer::rspVCont be independent of the last parsed "threadActions"
object.

 - We'll want to factor out that "wait or stops" code for use in other
   cases.

 - In non-stop mode, vCont and the "wait for stops" are decoupled.  We
   can see multiple vCont packets before any thread stops, even.

And for that, we need to keep track of what was the vCont last action
applied to each thread.

Since we're adding new flag to Thread, we might as well convert the
pendingStops set to a flag as well.  Since we'll normally have to
iterate the whole list of threads to check whether any halted, keeping
a separate list is questionable.  In any case, the current container
is a "std::set" (red black tree), sorted by thread id, which I very
much doubt is the best algorithm for this..  Thus, if we ever consider
optimizating the list of already-stopped-threads, I believe we'd do
something different anyway.