Rewrite/fix vCont support
Rework the vCont support in order to decouple the state a thread is in
from the vCont actions list.  Also replace the actions map with a
vector -- map doesn't work when we have actions like "pPID.-1" meaning
all threads of process PID, and also because the vCont actions should
be attempted from left to right until the first action matches.